requirements.txt
# 0DTE Options Trading Strategy Dependencies

# Data APIs
python-dotenv>=1.0.0
requests>=2.31.0

# Black-Scholes calculations
scipy>=1.11.0

# Alpaca Trading
alpaca-py>=0.26.0

# Scheduling
APScheduler>=3.10.0
pytz>=2024.1
